+++ lib/cocoapods/sandbox.rb
@@ -108,6 +108,12 @@ module Pod
       FileUtils.mkdir_p(target_support_files_root)
     end
 
+    # mark the install transaction is finished
+    def finish
+      FileUtils.rmtree install_locking_dir
+      @dirty_set = Set.new
+    end
+
     # @return [String] a string representation suitable for debugging.
     #
     def inspect
@@ -130,6 +136,11 @@ module Pod
     #
     def project_path
       root + 'Pods.xcodeproj'
+    end
+
+    # @return [Pathname] the path of the updating pods
+    def install_locking_dir
+      root + '.install_locking'
     end
 
     # @return [Pathname] the path of the installation cache.
@@ -442,6 +453,34 @@ module Pod
     def local_podspec(name)
       root_name = Specification.root_name(name)
       development_pods[root_name]
+    end
+
+    #--------------------------------------#
+
+    # mark the pod is dirty and need to be update
+    #
+    # @param [String] name
+    #        The name of the Pod
+    # @return [void]
+    def touch_pod(name)
+      root_name = Specification.root_name(name)
+      FileUtils.mkdir_p(install_locking_dir)
+      FileUtils.touch(install_locking_dir + root_name)
+      dirty_set.add(root_name)
+    end
+
+    # checks if a pod is dirty and need a overwrite
+    #
+    # @param [String] name
+    #        The name of the Pod
+    # @return [Bool] Whether the Pod is dirty
+    def dirty?(name)
+      root_name = Specification.root_name(name)
+      dirty_set.include?(root_name)
+    end
+
+    def dirty_set
+      @dirty_set ||= install_locking_dir.exist? ? Dir.children(install_locking_dir).to_set : Set.new
     end
 
     # @!group Convenience Methods